+++ test/smokes/code_sniffer/expectations.rb
@@ -8,7 +8,7 @@ s.add_test(
   issues: [
     {
       path: "app.php",
-      location: { start_line: 18 },
+      location: { start_line: 18, start_column: 1 },
       id: "PSR2.Files.ClosingTag.NotAllowed",
       message: "A closing tag is not permitted at the end of a PHP file",
       links: [],
@@ -25,7 +25,7 @@ s.add_test(
   issues: [
     {
       path: "app/app.php",
-      location: { start_line: 6 },
+      location: { start_line: 6, start_column: 1 },
       id: "PSR2.Files.ClosingTag.NotAllowed",
       message: "A closing tag is not permitted at the end of a PHP file",
       links: [],
@@ -61,40 +61,12 @@ s.add_test(
 )
 
 s.add_test(
-  "version_2",
-  type: "success",
-  issues: [
-    {
-      path: "app.php",
-      location: { start_line: 18 },
-      id: "PSR2.Files.ClosingTag.NotAllowed",
-      message: "A closing tag is not permitted at the end of a PHP file",
-      links: [],
-      object: { type: "ERROR", severity: 5, fixable: true },
-      git_blame_info: nil
-    }
-  ],
-  analyzer: { name: "PHP_CodeSniffer", version: "3.5.5" },
-  warnings: [
-    {
-      message: <<~MSG.strip,
-        DEPRECATION WARNING!!!
-        The following options in your `sider.yml` are deprecated and will be removed.
-        See https://help.sider.review/tools/php/code-sniffer for details.
-        - `linter.code_sniffer.version`
-      MSG
-      file: "sider.yml"
-    }
-  ]
-)
-
-s.add_test(
   "autodetect_cakephp",
   type: "success",
   issues: [
     {
       path: "app/test.php",
-      location: { start_line: 2 },
+      location: { start_line: 2, start_column: 1 },
       id: "CakePHP.Commenting.FunctionComment.Missing",
       message: "Missing doc comment for function foo()",
       links: [],
@@ -111,7 +83,7 @@ s.add_test(
   issues: [
     {
       path: "src/test.php",
-      location: { start_line: 2 },
+      location: { start_line: 2, start_column: 1 },
       id: "Squiz.Functions.GlobalFunction.Found",
       message: 'Consider putting global function "foo" in a static class',
       links: [],
@@ -120,7 +92,7 @@ s.add_test(
     },
     {
       path: "src/test.php",
-      location: { start_line: 2 },
+      location: { start_line: 2, start_column: 1 },
       id: "Symfony.Commenting.FunctionComment.Missing",
       message: "Missing doc comment for function foo()",
       links: [],