+++ test/smokes/detekt/expectations.rb
@@ -1,5 +1,7 @@
 s = Runners::Testing::Smoke
 
+default_version = "1.10.0"
+
 s.add_test(
   "with_broken_sider_yml",
   type: "failure",
@@ -9,7 +11,7 @@ s.add_test(
 
 s.add_test(
   "with_invalid_detekt_config",
-  type: "failure", analyzer: { name: "detekt", version: "1.9.1" }, message: "Your detekt configuration is invalid"
+  type: "failure", analyzer: { name: "detekt", version: default_version }, message: "Your detekt configuration is invalid"
 )
 
 s.add_test(
@@ -18,18 +20,18 @@ s.add_test(
   class: "RuntimeError",
   backtrace: :_,
   inspect: %r{.*'non/exists/dir/' does not exist.*},
-  analyzer: { name: "detekt", version: "1.9.1" }
+  analyzer: { name: "detekt", version: default_version }
 )
 
 s.add_test(
   "with_options",
   type: "success",
-  analyzer: { name: "detekt", version: "1.9.1" },
+  analyzer: { name: "detekt", version: default_version },
   issues: [
     {
       id: "detekt.EmptyClassBlock",
       path: "src/FilteredClass.kt",
-      location: { start_line: 2 },
+      location: { start_line: 2, start_column: 21 },
       message: "The class or object FilteredClass is empty.",
       links: [],
       object: { severity: "info" },
@@ -38,7 +40,7 @@ s.add_test(
     {
       id: "detekt.ForEachOnRange",
       path: "src/ComplexClass.kt",
-      location: { start_line: 44 },
+      location: { start_line: 44, start_column: 21 },
       message: "Using the forEach method on ranges has a heavy performance cost. Prefer using simple for loops.",
       links: [],
       object: { severity: "warning" },
@@ -47,7 +49,7 @@ s.add_test(
     {
       id: "detekt.FunctionOnlyReturningConstant",
       path: "src/App.kt",
-      location: { start_line: 8 },
+      location: { start_line: 8, start_column: 7 },
       message: "get is returning a constant. Prefer declaring a constant instead.",
       links: [],
       object: { severity: "warning" },
@@ -56,7 +58,7 @@ s.add_test(
     {
       id: "detekt.NestedBlockDepth",
       path: "src/ComplexClass.kt",
-      location: { start_line: 9 },
+      location: { start_line: 9, start_column: 13 },
       message: "Function complex is nested too deeply.",
       links: [],
       object: { severity: "warning" },
@@ -68,12 +70,12 @@ s.add_test(
 s.add_test(
   "without_options",
   type: "success",
-  analyzer: { name: "detekt", version: "1.9.1" },
+  analyzer: { name: "detekt", version: default_version },
   issues: [
     {
       id: "detekt.EmptyClassBlock",
       path: "src/FilteredClass.kt",
-      location: { start_line: 2 },
+      location: { start_line: 2, start_column: 21 },
       message: "The class or object FilteredClass is empty.",
       links: [],
       object: { severity: "info" },
@@ -82,7 +84,7 @@ s.add_test(
     {
       id: "detekt.ForEachOnRange",
       path: "src/ComplexClass.kt",
-      location: { start_line: 44 },
+      location: { start_line: 44, start_column: 21 },
       message: "Using the forEach method on ranges has a heavy performance cost. Prefer using simple for loops.",
       links: [],
       object: { severity: "warning" },
@@ -91,7 +93,7 @@ s.add_test(
     {
       id: "detekt.FunctionOnlyReturningConstant",
       path: "src/App.kt",
-      location: { start_line: 8 },
+      location: { start_line: 8, start_column: 7 },
       message: "get is returning a constant. Prefer declaring a constant instead.",
       links: [],
       object: { severity: "warning" },
@@ -100,7 +102,7 @@ s.add_test(
     {
       id: "detekt.MagicNumber",
       path: "src/ComplexClass.kt",
-      location: { start_line: 44 },
+      location: { start_line: 44, start_column: 17 },
       message: "This expression contains a magic number. Consider defining it to a well named constant.",
       links: [],
       object: { severity: "warning" },
@@ -109,7 +111,7 @@ s.add_test(
     {
       id: "detekt.MagicNumber",
       path: "src/ComplexClass.kt",
-      location: { start_line: 48 },
+      location: { start_line: 48, start_column: 26 },
       message: "This expression contains a magic number. Consider defining it to a well named constant.",
       links: [],
       object: { severity: "warning" },
@@ -118,7 +120,7 @@ s.add_test(
     {
       id: "detekt.NestedBlockDepth",
       path: "src/ComplexClass.kt",
-      location: { start_line: 9 },
+      location: { start_line: 9, start_column: 13 },
       message: "Function complex is nested too deeply.",
       links: [],
       object: { severity: "warning" },
@@ -127,7 +129,7 @@ s.add_test(
     {
       id: "detekt.TooGenericExceptionCaught",
       path: "src/ComplexClass.kt",
-      location: { start_line: 19 },
+      location: { start_line: 19, start_column: 22 },
       message:
         "Caught exception is too generic. Prefer catching specific exceptions to the case that is currently handled.",
       links: [],
@@ -137,7 +139,7 @@ s.add_test(
     {
       id: "detekt.TooGenericExceptionCaught",
       path: "src/ComplexClass.kt",
-      location: { start_line: 22 },
+      location: { start_line: 22, start_column: 26 },
       message:
         "Caught exception is too generic. Prefer catching specific exceptions to the case that is currently handled.",
       links: [],
@@ -147,7 +149,7 @@ s.add_test(
     {
       id: "detekt.TooGenericExceptionCaught",
       path: "src/ComplexClass.kt",
-      location: { start_line: 34 },
+      location: { start_line: 34, start_column: 26 },
       message:
         "Caught exception is too generic. Prefer catching specific exceptions to the case that is currently handled.",
       links: [],
@@ -160,12 +162,12 @@ s.add_test(
 s.add_test(
   "with_option_includes",
   type: "success",
-  analyzer: { name: "detekt", version: "1.9.1" },
+  analyzer: { name: "detekt", version: default_version },
   issues: [
     {
       id: "detekt.EmptyClassBlock",
       path: "src/main/App.kt",
-      location: { start_line: 1 },
+      location: { start_line: 1, start_column: 11 },
       message: "The class or object App is empty.",
       links: [],
       object: { severity: "info" },
@@ -174,4 +176,4 @@ s.add_test(
   ]
 )
 
-s.add_test("no_files", type: "success", analyzer: { name: "detekt", version: "1.9.1" }, issues: [])
+s.add_test("no_files", type: "success", analyzer: { name: "detekt", version: default_version }, issues: [])